requirements.txt
# requirements.txt
streamlit
google-generativeai
PyPDF2
nltk
pdf2image
Pillow
pytesseract